cnn fp8 inference with trt

+++ compressai/runtime/__init__.py
@@ -29,7 +29,7 @@ def build_runtime(net, codec: Any, config: RuntimeConfig):
             ga_runner,
             gs_runner,
             codec_input_dtype=torch.float32,
-            gs_input_dtype=torch.float32,     # torch fp32 baseline default
+            gs_input_dtype=torch.float16,     # torch fp32 baseline default
             ga_input_dtype=torch.float32,
         )
         return engine
@@ -57,7 +57,7 @@ def build_runtime(net, codec: Any, config: RuntimeConfig):
             gs_runner,
             codec_input_dtype=torch.float32,  # per your rule: codec always FP32
             gs_input_dtype=torch.float16,     # gs engine boundary dtype
-            ga_input_dtype=torch.float16,     # ga engine boundary dtype
+            ga_input_dtype=torch.float32,     # ga engine boundary dtype
         )
         return engine
